+++ examples/example_reports_actuals.py
@@ -32,7 +32,7 @@
 #  @author    Jeff Tanner <jefft@tune.com>
 #  @copyright 2014 Tune (http://www.tune.com)
 #  @license   http://opensource.org/licenses/MIT The MIT License (MIT)
-#  @version   0.9.13
+#  @version   $Date: 2014-11-03 15:19:08 $
 #  @link      https://developers.mobileapptracking.com @endlink
 #
 #  The Actuals report gives you quick insight into the performance of your apps
@@ -54,7 +54,8 @@ try:
         Stats,
         ReportReaderCSV,
         ReportReaderJSON,
-        TUNE_FIELDS_RECOMMENDED
+        TUNE_FIELDS_RECOMMENDED,
+        TUNE_FIELDS_DEFAULT
         )
 except ImportError as exc:
     sys.stderr.write("Error: failed to import module ({})".format(exc))
@@ -92,7 +93,16 @@ class ExampleReportsActuals(object):
 
             print("")
             print("======================================================")
-            print(" Fields of Advertiser Actuals records.       ")
+            print(" Fields of Advertiser Actuals records - Default.      ")
+            print("======================================================")
+
+            response = stats.fields(TUNE_FIELDS_DEFAULT)
+            for field in response:
+                print(str(field))
+
+            print("")
+            print("======================================================")
+            print(" Fields of Advertiser Actuals records - Recommended.  ")
             print("======================================================")
 
             response = stats.fields(TUNE_FIELDS_RECOMMENDED)
@@ -122,7 +132,31 @@ class ExampleReportsActuals(object):
 
             print("")
             print("======================================================")
-            print(" Find Advertiser Actuals records.            ")
+            print(" Find Advertiser Actuals records - Default fields.    ")
+            print("======================================================")
+
+            response = stats.find(
+                start_date,
+                end_date,
+                filter="(publisher_id > 0)",
+                fields=None,
+                limit=5,
+                page=None,
+                sort={"installs": "DESC"},
+                group="site_id,publisher_id",
+                timestamp="datehour",
+                response_timezone="America/Los_Angeles"
+            )
+
+            print("= TuneManagementResponse:")
+            print(str(response))
+
+            if response.http_code != 200:
+                raise Exception("Failed: {}: {}".format(response.http_code, str(response.errors)))
+
+            print("")
+            print("======================================================")
+            print(" Find Advertiser Actuals records - Recommended fields.")
             print("======================================================")
 
             response = stats.find(
@@ -132,7 +166,7 @@ class ExampleReportsActuals(object):
                 fields=stats.fields(TUNE_FIELDS_RECOMMENDED),
                 limit=5,
                 page=None,
-                sort={"site.name": "DESC"},
+                sort={"installs": "DESC"},
                 group="site_id,publisher_id",
                 timestamp="datehour",
                 response_timezone="America/Los_Angeles"