+++ tune/management/api/advertiser/stats/__init__.py
@@ -36,7 +36,7 @@ Tune Mangement API '/advertiser/stats/'
 #  @author    Jeff Tanner <jefft@tune.com>
 #  @copyright 2014 Tune (http://www.tune.com)
 #  @license   http://opensource.org/licenses/MIT The MIT License (MIT)
-#  @version   0.9.13
+#  @version   $Date: 2014-11-03 15:19:08 $
 #  @link      https://developers.mobileapptracking.com @endlink
 #